setting IDs to long

+++ src/main/java/com/rocs/infirmary/application/app/facade/medicine/inventory/MedicineInventoryFacade.java
@@ -41,7 +41,7 @@ public interface MedicineInventoryFacade {
      * @param expirationDate is used to specify the medicine expiration date
      * @return true if the inventory was successfully updated; false otherwise
      **/
-    boolean addInventory(int medicineId , String itemType, int quantity,Date expirationDate );
+    boolean addInventory(long medicineId , String itemType, int quantity, Date expirationDate );
     /**
      * This display all the medicine retrieved From the Medicine Table
      * @return a List of Medicine that is available on the database
@@ -52,7 +52,7 @@ public interface MedicineInventoryFacade {
      * @param inventoryId the ID of the inventory item to be deleted
      * @return true if the inventory record was successfully deleted false if not
      **/
-    boolean deleteInventory(int inventoryId);
+    boolean deleteInventory(long inventoryId);
     /**
      * this is used to update the medicine attributes in the inventory
      * @param inventoryId the ID of the inventory item to be updated
@@ -62,12 +62,12 @@ public interface MedicineInventoryFacade {
      * @param expirationDate is used to specify the medicine expiration date
      * @return true if the medicine is successfully updated and false if not
      **/
-    boolean updateMedicineInventory(int inventoryId,int medicineId, int quantity, String itemType, Date expirationDate);
+    boolean updateMedicineInventory(long inventoryId, long medicineId, int quantity, String itemType, Date expirationDate);
     /**
      * this is used to update the medicine attributes in the inventory
      * @param medicineId is a unique identifier that is used to specify the item to be updated
      * @param description is used to specify what type of item are going to be added on the item table
      * @return true if the medicine is successfully updated and false if not
      **/
-    boolean updateMedicine(int medicineId, String medicineName, String description);
+    boolean updateMedicine(long medicineId, String medicineName, String description);
 }
